core/requirements.txt
fastapi==0.75.1
mysql_connector_repackaged==0.3.1
quickchart.io==1.0.1
requests==2.25.1
uvicorn==0.17.6
python-multipart==0.0.5
gunicorn==20.1.0